Fix: Move WaitForSignal to base, so it can use used for testing other modules.
